+++ CHANGELOG.md
@@ -18,6 +18,7 @@
     - Random Forest now accepts a base estimator instance
     - CARTs now use max features heuristic by default
     - Added build/quick factory methods to Datasets
+    - Added Interval Discretizer transformer
 
 - 0.0.5-alpha
     - Added Gaussian Mixture clusterer